Avoid skipped live migration tests
This change I62006b7e636b1773f4f41709cfec1fd6daae3c4f (Oct 12, 2015)
modified the way in which the number of compute nodes is validated as a
requirement for the live migration tests. Instead of computing the number
of available compute nodes, a config option is used instead.

This patch sets this config option to the right value,  allowing live migration
tests to be run.